Added pascal voc 12 reference and refactored imagenet
