+++ examples/references/classification/imagenet/experiments/conda.yaml
@@ -1,0 +1,16 @@
+name: imagenet_env
+channels:
+  - defaults
+  - pytorch
+dependencies:
+  - python=3.7
+  - numpy
+  - pytorch>=1.1.0
+  - torchvision
+  - pip
+  - pip:
+    - mlflow
+    - tqdm
+    - tensorboardX
+    - git+https://github.com/vfdev-5/py_config_runner
+    - git+https://github.com/vfdev-5/ignite.git@distrib
