--- examples/references/imagenet/experiments/MLproject
@@ -1,12 +1,0 @@
-name: ImageNet
-
-conda_env: conda.yaml
-
-entry_points:
-
-  main:
-    parameters:
-      config_path: path
-      num_gpus: float
-    command: "sh setup_apex.sh && python -m torch.distributed.launch --nproc_per_node={num_gpus} `py_config_runner_script` ../code/training.py {config_path} --manual_config_load"
-
