+++ examples/references/segmentation/pascal_voc2012/.gitignore
@@ -1,0 +1,2 @@
+.ipynb_checkpoints
+output
